requirements.txt
langchain==0.3.25
langchain-core==0.3.63
langchain-community==0.3.24
langchain-openai==0.3.18
langchain-anthropic==0.3.7
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic-settings==2.6.1
pydantic>=2,<3
python-dotenv==1.1.0
openai>=1.68.2
requests==2.32.3
pandas>=2.0.0
numpy>=1.26.0,<2.0
pillow>=10.0.0
aiofiles>=23.0.0
aiohttp>=3.8.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
redis>=5.0.0
streamlit>=1.30.0
streamlit-lottie>=0.0.5
pytest>=8.0.0
pytest-asyncio>=0.23.0 